backend/auth-service/requirements.txt
fastapi
uvicorn
email-validator
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt
redis
httpx
python-dotenv
loguru
